requirements.txt
psycopg2-binary==2.9.9
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
jupyter==1.0.0
notebook==7.0.6
ipykernel==6.27.1
